agents/model_gateway.py
# model_gateway.py - Python version of ModelGateway
from typing import Any, Dict, Optional, Union

from typing_extensions import TypedDict


class ProviderConfig(TypedDict):
    models: list[str]
    api_key: Optional[str]


class RouterConfig(TypedDict):
    default_provider: str
    providers: Dict[str, ProviderConfig]


class ModelResponse(TypedDict):
    text: str
    usage: Dict[str, int]
    model: str
    provider: str


class ModelGateway:
    def __init__(self, router: RouterConfig) -> None:
        self.router = router

    def route_request(
        self, prompt: str, requirements: Optional[Dict[str, Any]] = None
    ) -> Optional[ModelResponse]:
        requirements = requirements or {}
        provider: str = requirements.get("provider") or self.router["default_provider"]
        provider_config = self.router["providers"].get(provider)
        if not provider_config:
            raise ValueError(f"Provider {provider} not configured")
        model: str = requirements.get("model") or provider_config["models"][0]
        if model not in provider_config["models"]:
            raise ValueError(f"Model {model} not available for provider {provider}")
        # TODO: Implement actual API calls to different providers
        if provider == "openai":
            return self.call_openai(prompt, model, requirements)
        elif provider == "anthropic":
            return self.call_anthropic(prompt, model, requirements)
        elif provider == "mistral":
            return self.call_mistral(prompt, model, requirements)
        else:
            raise ValueError(f"Unknown provider: {provider}")

    def call_openai(
        self, prompt: str, model: str, requirements: Dict[str, Any]
    ) -> Optional[ModelResponse]:
        # TODO: Implement OpenAI API call
        print(f"Calling OpenAI: {model}", prompt, requirements)
        return None

    def call_anthropic(
        self, prompt: str, model: str, requirements: Dict[str, Any]
    ) -> Optional[ModelResponse]:
        # TODO: Implement Anthropic API call
        print(f"Calling Anthropic: {model}", prompt, requirements)
        return None

    def call_mistral(
        self, prompt: str, model: str, requirements: Dict[str, Any]
    ) -> Optional[ModelResponse]:
        # TODO: Implement Mistral API call
        print(f"Calling Mistral: {model}", prompt, requirements)
        return None
