requirements.docker.txt
# Comprehensive requirements.txt for Cloudless.gr Platform
# This file contains all Python dependencies used across the platform
# including Jenkins CI/CD, testing, monitoring, and data processing

# Core Dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Testing Framework
pytest>=7.4.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0
pytest-cov>=4.1.0
pytest-json-report>=1.5.0
unittest-xml-reporting>=3.2.0
coverage>=7.3.0

# Code Quality & Security
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
safety>=2.3.0
bandit>=1.7.5
isort>=5.12.0
pylint>=3.0.0

# Data Science & Analytics
jupyter>=1.0.0
pandas>=2.1.0
numpy>=1.25.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
plotly>=5.17.0

# Development Tools
ipython>=8.16.0
notebook>=7.0.0
jupyterlab>=4.0.0

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
sqlmodel
asyncpg

# Redis & Caching
redis>=5.0.0
celery>=5.3.0

# Cloud & Storage
boto3>=1.29.0
minio>=7.2.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0
loguru>=0.7.0

# HTTP & API clients
httpx>=0.25.0
aiohttp>=3.9.0

# Configuration & Environment
click>=8.1.0
pyyaml>=6.0.1
toml>=0.10.2
configparser>=6.0.0

# Date & Time
arrow>=1.3.0
pytz>=2023.3

# Image Processing (optional, for future use)
Pillow>=10.1.0

# Email & Notifications
sendgrid>=6.10.0
twilio>=8.10.0

# Machine Learning (using Ollama as LLM backend)
# No additional ML dependencies required

# Development & Debugging
rich>=13.7.0
typer>=0.9.0
pdbpp>=0.10.3

# Async & Concurrency
asyncio-mqtt>=0.16.0
aiofiles>=23.2.1

# Validation & Serialization
marshmallow>=3.20.0
cerberus>=1.3.4

# PDF & Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0

# Export & Reporting
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Cryptography & Security
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Web Scraping (optional)
beautifulsoup4>=4.12.0
scrapy>=2.11.0

# Task Queue
rq>=1.15.0

# API Documentation
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0

# Performance & Profiling
py-spy>=0.3.14
memory-profiler>=0.61.0

# Additional utilities
python-magic>=0.4.27
chardet>=5.2.0
tqdm>=4.66.0
